pyproject.toml
[project]
name = "twilio-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Moritz Enderle",email = "contact@moritzenderle.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "twilio (>=9.8.3,<10.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "xai-sdk (>=1.6.1,<2.0.0)",
    "openai (>=1.0.0,<2.0.0)",
    "matplotlib (>=3.8.0,<4.0.0)",
    "networkx (>=3.2.0,<4.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "isort (>=6.1.0,<7.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "python-levenshtein (>=0.27.1,<0.28.0)",
    "google-api-core (>=2.25.1,<3.0.0)",
    "google-maps-routing (>=0.6.16,<0.7.0)",
    "pytz (>=2025.2,<2026.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "tzdata (>=2025.2,<2026.0)",
    "python-telegram-bot (>=22.5,<23.0)",
    "requests (>=2.25.0,<3.0.0)",
    "elevenlabs (>=2.20.1,<3.0.0)",
    "num2words (>=0.5.14,<0.6.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "strict"
filterwarnings = [
    "ignore::RuntimeWarning:unittest.mock",
]

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"

